ingestion/requirements.txt
paho-mqtt
psycopg2-binary
pytest
python-dotenv